helper-apps/cortex-file-converter/requirements.txt
azure-storage-blob 